added autostart.bat and autostart tutorial
added autostart.bat which hopefully works. In that case it would create a shortcut to the script in the autostart folder, so that the script is automatically run when you start the pc (when you log in).
Added Tutorial on how to do that /|\
(how to start autostart.bat basically)